Fix speaker autoresearch promotion safety
