+++ .agents/test-matrix.yml
@@ -388,6 +388,9 @@ rules:
       - "scripts/score_speaker_eval.py"
       - "scripts/aggregate_sweep.py"
       - "scripts/run_speaker_autoresearch.py"
+      - "scripts/speaker_autoresearch_contract.py"
+      - "scripts/speaker_autoresearch_runtime.py"
+      - "scripts/test_speaker_autoresearch.py"
     checks:
       - "scripts/dev/agent-preflight.sh"
       - "bash -n scripts/download_ami.sh"
@@ -395,6 +398,9 @@ rules:
       - "python3 -m py_compile scripts/score_speaker_eval.py"
       - "python3 -m py_compile scripts/aggregate_sweep.py"
       - "python3 -m py_compile scripts/run_speaker_autoresearch.py"
+      - "python3 -m py_compile scripts/speaker_autoresearch_contract.py"
+      - "python3 -m py_compile scripts/speaker_autoresearch_runtime.py"
+      - "python3 -m py_compile scripts/test_speaker_autoresearch.py"
       - "python3 scripts/test_speaker_autoresearch.py"
       - "bash build-deps.sh --force"
       - "swift build --package-path Tools/SpeakerEvalHarness"
