+++ Tools/SpeakerEvalHarness/Sources/speaker-eval-harness/AutoResearch.swift
@@ -2,303 +2,6 @@ import Foundation
 import CryptoKit
 import TranscriptedCore
 
-// MARK: - Frozen fingerprint input
-
-private struct FingerprintSpeaker: Decodable {
-    let gtSpeaker: String
-    let embedding: [Float]
-    let durationSeconds: Double
-    let segmentCount: Int
-    let clusterCount: Int
-    let purity: Double
-}
-
-private struct FingerprintMeeting: Decodable {
-    let meeting: String
-    let order: Int
-    let speakers: [FingerprintSpeaker]
-}
-
-private struct FingerprintCache: Decodable {
-    let corpus: String
-    let meetings: [FingerprintMeeting]
-}
-
-private enum ResearchSplit: String, Codable, CaseIterable {
-    case train
-    case dev
-    case holdout
-    case all
-}
-
-private enum MaturityEvidence: String, Codable {
-    case appearances
-    case confirmedMeetings = "confirmed_meetings"
-}
-
-private enum WriteBackEvidence: String, Codable {
-    /// Current production behavior: every accepted DB match records an appearance and may blend.
-    case production
-    /// Defer suggest-path blending until the user confirms it; auto-name still adapts.
-    case confirmedOrAuto = "confirmed_or_auto"
-    /// Only an explicit user confirmation may change the voiceprint.
-    case confirmedOnly = "confirmed_only"
-}
-
-private struct AutoResearchConfig: Codable, Equatable {
-    let id: String
-    let autoMaturityEvidence: MaturityEvidence
-    let matchMaturityEvidence: MaturityEvidence
-    let requiredMaturityCount: Int
-    let autoSimilarity: Double
-    let autoMargin: Double
-    let minimumAverageSimilarity: Double
-    let minimumSpeechSeconds: Double
-    let minimumSegmentCount: Int
-    let matchFloorOffset: Double
-    let writeBackEvidence: WriteBackEvidence
-    let writeBackMargin: Double
-    let confidentWriteSimilarity: Double
-    let cautiousWriteSimilarity: Double
-    let confidentBlendAlpha: Float
-    let cautiousBlendAlpha: Float
-    let maximumExemplars: Int
-    let exemplarSameConditionSimilarity: Double
-    let exemplarBlendAlpha: Float
-
-    static let productionBaseline = AutoResearchConfig(
-        id: "baseline-production",
-        autoMaturityEvidence: .confirmedMeetings,
-        matchMaturityEvidence: .appearances,
-        requiredMaturityCount: SpeakerNamingPolicy.requiredConfirmedMeetings,
-        autoSimilarity: SpeakerNamingPolicy.autoAcceptSimilarityThreshold,
-        autoMargin: SpeakerNamingPolicy.autoAcceptMarginMin,
-        minimumAverageSimilarity: -1,
-        minimumSpeechSeconds: 0,
-        minimumSegmentCount: 0,
-        matchFloorOffset: 0,
-        writeBackEvidence: .production,
-        writeBackMargin: SpeakerWritePathPolicy.writeBackMarginMin,
-        confidentWriteSimilarity: SpeakerWritePathPolicy.confidentWriteBackSimilarity,
-        cautiousWriteSimilarity: SpeakerWritePathPolicy.cautiousWriteBackSimilarity,
-        confidentBlendAlpha: SpeakerWritePathPolicy.confidentBlendAlpha,
-        cautiousBlendAlpha: SpeakerWritePathPolicy.cautiousBlendAlpha,
-        maximumExemplars: SpeakerExemplarPolicy.maxExemplars,
-        exemplarSameConditionSimilarity: SpeakerExemplarPolicy.sameConditionSimilarity,
-        exemplarBlendAlpha: SpeakerExemplarPolicy.exemplarBlendAlpha
-    )
-}
-
-private struct ConfigEnvelope: Decodable {
-    let configs: [AutoResearchConfig]
-}
-
-// MARK: - Report schema
-
-private struct MetricSnapshot: Codable {
-    let observations: Int
-    let scorableObservations: Int
-    let recurringSpeakerUnits: Int
-    let returningOpportunities: Int
-    let asks: Int
-    let suggestions: Int
-    let automaticNames: Int
-    let correctAutomaticNames: Int
-    let falseAutomaticNames: Int
-    let falseAutomaticNamesAllPurities: Int
-    let automaticNamesOnLowPurity: Int
-    let wrongSuggestions: Int
-    let repeatPrompts: Int
-    let openSetTrials: Int
-    let openSetFalseAutomaticNames: Int
-    let openSetWrongSuggestions: Int
-    let falseMergeIndicators: Int
-    let withinMeetingFalseMergeIndicators: Int
-    let crossMeetingFalseMergeIndicators: Int
-    let fragmentationExcess: Int
-    let contaminatedProfiles: Int
-    let profilesAtEnd: Int
-    let identitiesReachingAuto: Int
-    let meanAppearanceToFirstAuto: Double?
-    let meanConfirmedMeetingsAtFirstAuto: Double?
-    let promptsPerRecurringSpeaker: Double?
-    let autoCoverage: Double?
-    let autoPrecision: Double?
-    let falseAutoUpper95WhenZero: Double?
-}
-
-private struct TraceEvent: Codable {
-    let corpus: String
-    let meeting: String
-    let meetingOrder: Int
-    let truthSpeaker: String
-    let purity: Double
-    let speechSeconds: Double
-    let segmentCount: Int
-    let clusterCount: Int
-    let decision: String
-    let predictedProfileId: UUID?
-    let predictedTruth: String?
-    let resolvedProfileId: UUID
-    let similarity: Double?
-    let secondBestSimilarity: Double?
-    let averageSimilarity: Double?
-    let secondBestAverageSimilarity: Double?
-    let matchThreshold: Double
-    let appearanceCountAtDecision: Int?
-    let confirmedMeetingsAtDecision: Int?
-    let correct: Bool?
-}
-
-private struct ConfigReport: Codable {
-    let config: AutoResearchConfig
-    let metrics: MetricSnapshot
-    let slices: [String: MetricSnapshot]
-    let events: [TraceEvent]?
-}
-
-private struct AutoResearchReport: Codable {
-    let schemaVersion: Int
-    let generatedAt: String
-    let split: ResearchSplit
-    let manifestPath: String
-    let inputRoot: String
-    let scoringPurityFloor: Double
-    let splitContract: String
-    let reports: [ConfigReport]
-}
-
-// MARK: - Mutable simulation state
-
-private struct SimulatedProfile {
-    var value: SpeakerProfile
-    let labelTruth: String
-    var appearanceCount: Int
-    var confirmedMeetings: Set<String>
-    /// Most-recent-first, matching `SpeakerDatabase.recentMatchOutcomes`.
-    var recentOutcomes: [SpeakerMatchOutcomeKind]
-    var negativeExemplars: [[Float]]
-    var exemplarCounts: [Int]
-    var observedTruths: Set<String>
-}
-
-private struct MutableMetrics {
-    var observations = 0
-    var scorableObservations = 0
-    var recurringSpeakerUnits = 0
-    var returningOpportunities = 0
-    var asks = 0
-    var suggestions = 0
-    var automaticNames = 0
-    var correctAutomaticNames = 0
-    var falseAutomaticNames = 0
-    var falseAutomaticNamesAllPurities = 0
-    var automaticNamesOnLowPurity = 0
-    var wrongSuggestions = 0
-    var repeatPrompts = 0
-    var openSetTrials = 0
-    var openSetFalseAutomaticNames = 0
-    var openSetWrongSuggestions = 0
-    var withinMeetingFalseMergeIndicators = 0
-    var crossMeetingFalseMergeIndicators = 0
-    var fragmentationExcess = 0
-    var contaminatedProfiles = 0
-    var profilesAtEnd = 0
-    var firstAutoAppearanceTotal = 0
-    var firstAutoConfirmedTotal = 0
-    var identitiesReachingAuto = 0
-
-    mutating func merge(_ other: MutableMetrics) {
-        observations += other.observations
-        scorableObservations += other.scorableObservations
-        recurringSpeakerUnits += other.recurringSpeakerUnits
-        returningOpportunities += other.returningOpportunities
-        asks += other.asks
-        suggestions += other.suggestions
-        automaticNames += other.automaticNames
-        correctAutomaticNames += other.correctAutomaticNames
-        falseAutomaticNames += other.falseAutomaticNames
-        falseAutomaticNamesAllPurities += other.falseAutomaticNamesAllPurities
-        automaticNamesOnLowPurity += other.automaticNamesOnLowPurity
-        wrongSuggestions += other.wrongSuggestions
-        repeatPrompts += other.repeatPrompts
-        openSetTrials += other.openSetTrials
-        openSetFalseAutomaticNames += other.openSetFalseAutomaticNames
-        openSetWrongSuggestions += other.openSetWrongSuggestions
-        withinMeetingFalseMergeIndicators += other.withinMeetingFalseMergeIndicators
-        crossMeetingFalseMergeIndicators += other.crossMeetingFalseMergeIndicators
-        fragmentationExcess += other.fragmentationExcess
-        contaminatedProfiles += other.contaminatedProfiles
-        profilesAtEnd += other.profilesAtEnd
-        firstAutoAppearanceTotal += other.firstAutoAppearanceTotal
-        firstAutoConfirmedTotal += other.firstAutoConfirmedTotal
-        identitiesReachingAuto += other.identitiesReachingAuto
-    }
-
-    func snapshot() -> MetricSnapshot {
-        let promptsPerSpeaker = recurringSpeakerUnits > 0
-            ? Double(repeatPrompts) / Double(recurringSpeakerUnits) : nil
-        let coverage = returningOpportunities > 0
-            ? Double(correctAutomaticNames) / Double(returningOpportunities) : nil
-        let precision = automaticNames > 0
-            ? Double(automaticNames - falseAutomaticNames) / Double(automaticNames) : nil
-        let firstAuto = identitiesReachingAuto > 0
-            ? Double(firstAutoAppearanceTotal) / Double(identitiesReachingAuto) : nil
-        let firstAutoConfirmed = identitiesReachingAuto > 0
-            ? Double(firstAutoConfirmedTotal) / Double(identitiesReachingAuto) : nil
-        let upper = falseAutomaticNames == 0 && automaticNames > 0
-            ? min(1, 3.0 / Double(automaticNames)) : nil
-        return MetricSnapshot(
-            observations: observations,
-            scorableObservations: scorableObservations,
-            recurringSpeakerUnits: recurringSpeakerUnits,
-            returningOpportunities: returningOpportunities,
-            asks: asks,
-            suggestions: suggestions,
-            automaticNames: automaticNames,
-            correctAutomaticNames: correctAutomaticNames,
-            falseAutomaticNames: falseAutomaticNames,
-            falseAutomaticNamesAllPurities: falseAutomaticNamesAllPurities,
-            automaticNamesOnLowPurity: automaticNamesOnLowPurity,
-            wrongSuggestions: wrongSuggestions,
-            repeatPrompts: repeatPrompts,
-            openSetTrials: openSetTrials,
-            openSetFalseAutomaticNames: openSetFalseAutomaticNames,
-            openSetWrongSuggestions: openSetWrongSuggestions,
-            falseMergeIndicators: withinMeetingFalseMergeIndicators + crossMeetingFalseMergeIndicators,
-            withinMeetingFalseMergeIndicators: withinMeetingFalseMergeIndicators,
-            crossMeetingFalseMergeIndicators: crossMeetingFalseMergeIndicators,
-            fragmentationExcess: fragmentationExcess,
-            contaminatedProfiles: contaminatedProfiles,
-            profilesAtEnd: profilesAtEnd,
-            identitiesReachingAuto: identitiesReachingAuto,
-            meanAppearanceToFirstAuto: firstAuto,
-            meanConfirmedMeetingsAtFirstAuto: firstAutoConfirmed,
-            promptsPerRecurringSpeaker: promptsPerSpeaker,
-            autoCoverage: coverage,
-            autoPrecision: precision,
-            falseAutoUpper95WhenZero: upper
-        )
-    }
-}
-
-private struct PendingObservation {
-    let input: FingerprintSpeaker
-    var match: Transcription.SnapshotMatchResult?
-    let matchThreshold: Double
-}
-
-private struct CacheEvaluation {
-    var metrics: MutableMetrics
-    var events: [TraceEvent]
-}
-
-private struct AccumulatedConfig {
-    var metrics = MutableMetrics()
-    var slices: [String: MutableMetrics] = [:]
-    var events: [TraceEvent] = []
-}
-
 // MARK: - Command
 
 func runAutoResearch(_ args: [String]) {
@@ -333,6 +36,19 @@ func runAutoResearch(_ args: [String]) {
             let result = evaluate(cache: cache, config: config, split: split, includeEvents: includeEvents)
             accumulated[config.id]!.metrics.merge(result.metrics)
             accumulated[config.id]!.slices[cache.corpus, default: MutableMetrics()].merge(result.metrics)
+            for condition in ResearchConditionSlice.allCases {
+                let conditionResult = evaluate(
+                    cache: cache,
+                    config: config,
+                    split: split,
+                    includeEvents: false,
+                    condition: condition
+                )
+                accumulated[config.id]!.slices[
+                    condition.rawValue,
+                    default: MutableMetrics()
+                ].merge(conditionResult.metrics)
+            }
             if includeEvents { accumulated[config.id]!.events.append(contentsOf: result.events) }
         }
         FileHandle.standardError.write(Data("[autoeval] loaded \(cache.corpus) (\(cache.meetings.count) meetings)\n".utf8))
@@ -348,7 +64,7 @@ func runAutoResearch(_ args: [String]) {
         )
     }
     let report = AutoResearchReport(
-        schemaVersion: 2,
+        schemaVersion: 3,
         generatedAt: ISO8601DateFormatter().string(from: Date()),
         split: split,
         manifestPath: URL(fileURLWithPath: manifest).standardizedFileURL.path,
@@ -452,7 +168,7 @@ private func loadManifest(_ path: String, inputRoot: String) -> [String] {
     return inputs
 }
 
-private func sha256Hex(of url: URL) throws -> String {
+func sha256Hex(of url: URL) throws -> String {
     let handle = try FileHandle(forReadingFrom: url)
     defer { try? handle.close() }
     var hasher = SHA256()
@@ -464,28 +180,32 @@ private func sha256Hex(of url: URL) throws -> String {
 
 // MARK: - Chronological replay
 
-private enum ResearchConstants {
+enum ResearchConstants {
     static let scoringPurityFloor = 0.80
 }
 
-private func evaluate(
+func evaluate(
     cache: FingerprintCache,
     config: AutoResearchConfig,
     split: ResearchSplit,
-    includeEvents: Bool
+    includeEvents: Bool,
+    condition: ResearchConditionSlice? = nil
 ) -> CacheEvaluation {
     let family = corpusFamily(cache.corpus)
     let orderedMeetings = cache.meetings.sorted {
         ($0.order, $0.meeting) < ($1.order, $1.meeting)
     }
-    let filteredMeetings = orderedMeetings.map { meeting in
-        FingerprintMeeting(
-            meeting: meeting.meeting,
-            order: meeting.order,
-            speakers: meeting.speakers.filter { split == .all || identitySplit(family: family, truth: $0.gtSpeaker) == split }
-        )
-    }
-    let appearances = Dictionary(grouping: filteredMeetings.flatMap(\.speakers), by: \.gtSpeaker).mapValues(\.count)
+    let targetTruths = Set(orderedMeetings.flatMap(\.speakers).compactMap { speaker in
+        split == .all || identitySplit(family: family, truth: speaker.gtSpeaker) == split
+            ? speaker.gtSpeaker
+            : nil
+    })
+    let appearances = Dictionary(
+        grouping: orderedMeetings.flatMap(\.speakers).filter {
+            targetTruths.contains($0.gtSpeaker) && (condition?.contains($0) ?? true)
+        },
+        by: \.gtSpeaker
+    ).mapValues(\.count)
 
     var metrics = MutableMetrics()
     metrics.recurringSpeakerUnits = appearances.values.filter { $0 >= 2 }.count
@@ -496,11 +216,15 @@ private func evaluate(
     var identitiesWithFirstAuto: Set<String> = []
     var predictedProfilesByTruth: [String: Set<UUID>] = [:]
     var predictedMeetingsByProfileAndTruth: [UUID: [String: Set<String>]] = [:]
+    var scoredMeetingsByProfileAndTruth: [UUID: [String: Set<String>]] = [:]
+    var scoredMutationProfileIds: Set<UUID> = []
     var events: [TraceEvent] = []
     var nextProfileNumber: UInt64 = 1
 
-    for meeting in filteredMeetings where !meeting.speakers.isEmpty {
+    for meeting in orderedMeetings where !meeting.speakers.isEmpty {
         let meetingKey = "\(cache.corpus):\(meeting.meeting)"
+        let seenTruthsAtMeetingStart = seenTruths
+        let preMeetingProfiles = profiles
         let snapshotIds = profiles.keys.sorted { $0.uuidString < $1.uuidString }
         let snapshotProfiles = snapshotIds.compactMap { id -> SpeakerProfile? in
             guard var profile = profiles[id]?.value, let state = profiles[id] else { return nil }
@@ -557,8 +281,10 @@ private func evaluate(
             return cursor
         }
         let spunOff = Set(linkPlan.spinOffs)
+        var representativeByIndex: [Int: Int] = [:]
         for i in pending.indices {
             let rep = representative(for: i)
+            representativeByIndex[i] = rep
             if spunOff.contains(rep) {
                 pending[i].match = nil
             } else if rep != i {
@@ -566,18 +292,31 @@ private func evaluate(
             }
         }
 
-        for observation in pending {
+        let representativeIndices = pending.indices.filter { representativeByIndex[$0] == $0 }
+        let memberIndices = pending.indices.filter { representativeByIndex[$0] != $0 }
+        var resolvedIdByRepresentative: [Int: UUID] = [:]
+        for index in representativeIndices + memberIndices {
+            let observation = pending[index]
+            let representativeIndex = representativeByIndex[index] ?? index
+            let isRepresentative = representativeIndex == index
             let speaker = observation.input
-            let wasSeen = seenTruths.contains(speaker.gtSpeaker)
+            let scoresObservation = targetTruths.contains(speaker.gtSpeaker)
+                && (condition?.contains(speaker) ?? true)
+            let wasSeen = seenTruthsAtMeetingStart.contains(speaker.gtSpeaker)
             let galleryWasNonEmpty = !snapshotProfiles.isEmpty
-            appearanceNumber[speaker.gtSpeaker, default: 0] += 1
-            let currentAppearance = appearanceNumber[speaker.gtSpeaker]!
-            metrics.observations += 1
-            if speaker.purity >= ResearchConstants.scoringPurityFloor { metrics.scorableObservations += 1 }
-            if wasSeen { metrics.returningOpportunities += 1 }
-            if !wasSeen && galleryWasNonEmpty { metrics.openSetTrials += 1 }
+            if scoresObservation {
+                appearanceNumber[speaker.gtSpeaker, default: 0] += 1
+                metrics.observations += 1
+                if speaker.purity >= ResearchConstants.scoringPurityFloor { metrics.scorableObservations += 1 }
+                if wasSeen { metrics.returningOpportunities += 1 }
+                if !wasSeen && galleryWasNonEmpty { metrics.openSetTrials += 1 }
+            }
+            let currentAppearance = appearanceNumber[speaker.gtSpeaker] ?? 0
 
-            let matchedState = observation.match.flatMap { profiles[$0.profileId] }
+            // Every decision in one meeting reads the same frozen profile state. Production
+            // matches and classifies against its pre-meeting snapshot, and fused member clusters
+            // do not get a second chance to advance maturity after their representative writes.
+            let matchedState = observation.match.flatMap { preMeetingProfiles[$0.profileId] }
             let predictedTruth = matchedState?.labelTruth
             let correct = predictedTruth.map { $0 == speaker.gtSpeaker }
             let decision: String
@@ -590,7 +329,7 @@ private func evaluate(
                 decision = "ask"
             }
 
-            switch decision {
+            if scoresObservation { switch decision {
             case "auto":
                 metrics.automaticNames += 1
                 if speaker.purity < ResearchConstants.scoringPurityFloor { metrics.automaticNamesOnLowPurity += 1 }
@@ -617,10 +356,16 @@ private func evaluate(
             default:
                 metrics.asks += 1
                 if wasSeen { metrics.repeatPrompts += 1 }
-            }
+            } }
 
             let resolvedId: UUID
-            if let match = observation.match, decision == "auto" {
+            if !isRepresentative,
+               let inherited = resolvedIdByRepresentative[representativeIndex] {
+                // Production removes remapped members from the write-back set. They inherit
+                // the representative's final identity but never add another appearance,
+                // confirmation, correction, or exemplar update.
+                resolvedId = inherited
+            } else if let match = observation.match, decision == "auto" {
                 resolvedId = match.profileId
                 applyMatchedAppearance(
                     profileId: match.profileId,
@@ -707,15 +452,29 @@ private func evaluate(
                     )
                 }
             }
+            if isRepresentative {
+                resolvedIdByRepresentative[index] = resolvedId
+                if scoresObservation {
+                    scoredMutationProfileIds.insert(resolvedId)
+                }
+            }
 
             let predictedId = observation.match?.profileId ?? resolvedId
-            predictedProfilesByTruth[speaker.gtSpeaker, default: []].insert(predictedId)
+            if scoresObservation {
+                predictedProfilesByTruth[speaker.gtSpeaker, default: []].insert(predictedId)
+            }
             predictedMeetingsByProfileAndTruth[predictedId, default: [:]][
                 speaker.gtSpeaker,
                 default: []
             ].insert(meetingKey)
+            if scoresObservation {
+                scoredMeetingsByProfileAndTruth[predictedId, default: [:]][
+                    speaker.gtSpeaker,
+                    default: []
+                ].insert(meetingKey)
+            }
 
-            if includeEvents {
+            if includeEvents && scoresObservation {
                 events.append(TraceEvent(
                     corpus: cache.corpus,
                     meeting: meeting.meeting,
@@ -744,9 +503,14 @@ private func evaluate(
     }
 
     metrics.fragmentationExcess = predictedProfilesByTruth.values.reduce(0) { $0 + max(0, $1.count - 1) }
-    metrics.contaminatedProfiles = profiles.values.filter { $0.observedTruths.count > 1 }.count
-    metrics.profilesAtEnd = profiles.count
-    let falseMerges = falseMergeIndicators(predictedMeetingsByProfileAndTruth)
+    metrics.contaminatedProfiles = profiles.values.filter {
+        $0.observedTruths.count > 1 && scoredMutationProfileIds.contains($0.value.id)
+    }.count
+    metrics.profilesAtEnd = Set(predictedProfilesByTruth.values.flatMap { $0 }).count
+    let falseMerges = falseMergeIndicators(
+        predictedMeetingsByProfileAndTruth,
+        scoredMeetingsByProfileAndTruth: scoredMeetingsByProfileAndTruth
+    )
     metrics.withinMeetingFalseMergeIndicators = falseMerges.withinMeeting
     metrics.crossMeetingFalseMergeIndicators = falseMerges.crossMeeting
     return CacheEvaluation(metrics: metrics, events: events)
@@ -756,17 +520,26 @@ private func evaluate(
 /// A pair is within-meeting when the two truths ever shared a meeting; otherwise
 /// it is cross-meeting. The buckets are intentionally disjoint so their sum is
 /// a meaningful total instead of counting one collision twice.
-private func falseMergeIndicators(
-    _ meetingsByProfileAndTruth: [UUID: [String: Set<String>]]
+func falseMergeIndicators(
+    _ meetingsByProfileAndTruth: [UUID: [String: Set<String>]],
+    scoredMeetingsByProfileAndTruth: [UUID: [String: Set<String>]]? = nil
 ) -> (withinMeeting: Int, crossMeeting: Int) {
     var withinMeeting = 0
     var crossMeeting = 0
 
-    for meetingsByTruth in meetingsByProfileAndTruth.values {
+    for (profileId, meetingsByTruth) in meetingsByProfileAndTruth {
         let truths = meetingsByTruth.keys.sorted()
         guard truths.count > 1 else { continue }
         for firstIndex in 0..<(truths.count - 1) {
             for secondIndex in (firstIndex + 1)..<truths.count {
+                if let scoredMeetingsByProfileAndTruth {
+                    let scoredByTruth = scoredMeetingsByProfileAndTruth[profileId] ?? [:]
+                    let firstWasScored = !(scoredByTruth[truths[firstIndex]]?.isEmpty ?? true)
+                    let secondWasScored = !(scoredByTruth[truths[secondIndex]]?.isEmpty ?? true)
+                    if !firstWasScored && !secondWasScored {
+                        continue
+                    }
+                }
                 let firstMeetings = meetingsByTruth[truths[firstIndex]] ?? []
                 let secondMeetings = meetingsByTruth[truths[secondIndex]] ?? []
                 if firstMeetings.isDisjoint(with: secondMeetings) {
@@ -780,7 +553,7 @@ private func falseMergeIndicators(
     return (withinMeeting, crossMeeting)
 }
 
-private func shouldAutoName(
+func shouldAutoName(
     state: SimulatedProfile,
     match: Transcription.SnapshotMatchResult,
     speaker: FingerprintSpeaker,
@@ -821,7 +594,7 @@ private func shouldAutoName(
     return decision
 }
 
-private func applyConfirmedSuggestion(
+func applyConfirmedSuggestion(
     profileId: UUID,
     speaker: FingerprintSpeaker,
     match: Transcription.SnapshotMatchResult,
@@ -895,7 +668,7 @@ private func applyUserConfirmedAppearance(
 /// temporary profile, then the user's known-person choice merges that profile
 /// into the existing keeper using call-count weights. Structural merges clear
 /// the keeper's exemplar cache and add the production confidence bump.
-private func applyUserConfirmedMerge(
+func applyUserConfirmedMerge(
     profileId: UUID,
     speaker: FingerprintSpeaker,
     profiles: inout [UUID: SimulatedProfile]
@@ -942,146 +715,7 @@ private func recordRecentOutcome(
     profiles[profileId] = state
 }
 
-/// Cheap parity checks run by the verification matrix. These exercise the two
-/// failure modes most likely to make an offline sweep disagree with production:
-/// correction probation and confirmation-driven repair.
-func runAutoResearchSelfTests() {
-    let profileId = deterministicProfileId(1)
-    let embedding: [Float] = [1, 0]
-    let profile = SpeakerProfile(
-        id: profileId,
-        displayName: "truth:test",
-        nameSource: NameSource.userManual,
-        embedding: embedding,
-        firstSeen: Date(timeIntervalSince1970: 1),
-        lastSeen: Date(timeIntervalSince1970: 1),
-        callCount: 8,
-        confidence: 0.9,
-        disputeCount: 0,
-        confirmedMeetingCount: SpeakerNamingPolicy.requiredConfirmedMeetings,
-        exemplars: []
-    )
-    let confirmedMeetings = Set(
-        (1...SpeakerNamingPolicy.requiredConfirmedMeetings).map { "meeting-\($0)" }
-    )
-    let speaker = FingerprintSpeaker(
-        gtSpeaker: "test",
-        embedding: embedding,
-        durationSeconds: 10,
-        segmentCount: 4,
-        clusterCount: 1,
-        purity: 1
-    )
-    let match = Transcription.SnapshotMatchResult(
-        profileId: profileId,
-        similarity: 0.99,
-        secondBestSimilarity: -1,
-        averageSimilarity: 0.99,
-        secondBestAverageSimilarity: -1
-    )
-    var correctedState = SimulatedProfile(
-        value: profile,
-        labelTruth: "test",
-        appearanceCount: 8,
-        confirmedMeetings: confirmedMeetings,
-        recentOutcomes: [.corrected],
-        negativeExemplars: [],
-        exemplarCounts: [],
-        observedTruths: ["test"]
-    )
-    guard !shouldAutoName(
-        state: correctedState,
-        match: match,
-        speaker: speaker,
-        config: .productionBaseline
-    ) else {
-        die("autoeval self-test failed: a recent correction did not force probation")
-    }
-
-    correctedState.recentOutcomes = [.confirmed, .corrected]
-    guard shouldAutoName(
-        state: correctedState,
-        match: match,
-        speaker: speaker,
-        config: .productionBaseline
-    ) else {
-        die("autoeval self-test failed: a later confirmation did not restore trust")
-    }
-
-    correctedState.value.disputeCount = 2
-    var profiles = [profileId: correctedState]
-    applyConfirmedSuggestion(
-        profileId: profileId,
-        speaker: speaker,
-        match: match,
-        meeting: "repair-meeting",
-        config: .productionBaseline,
-        profiles: &profiles
-    )
-    guard profiles[profileId]?.value.disputeCount == 0,
-          profiles[profileId]?.recentOutcomes.first == .confirmed else {
-        die("autoeval self-test failed: confirmation did not repair profile state")
-    }
-
-    var mergeState = correctedState
-    mergeState.value.confidence = 0.5
-    mergeState.value.exemplars = [[0, 1]]
-    mergeState.exemplarCounts = [2]
-    profiles = [profileId: mergeState]
-    let unmatchedSpeaker = FingerprintSpeaker(
-        gtSpeaker: "test",
-        embedding: [0, 1],
-        durationSeconds: 10,
-        segmentCount: 4,
-        clusterCount: 1,
-        purity: 1
-    )
-    applyUserConfirmedMerge(
-        profileId: profileId,
-        speaker: unmatchedSpeaker,
-        profiles: &profiles
-    )
-    let expectedMergeEmbedding = blend(embedding, unmatchedSpeaker.embedding, alpha: 1 / 9)
-    guard let merged = profiles[profileId],
-          merged.appearanceCount == 9,
-          abs(merged.value.confidence - 0.65) < 0.0001,
-          merged.value.exemplars.isEmpty,
-          merged.exemplarCounts.isEmpty,
-          zip(merged.value.embedding, expectedMergeEmbedding).allSatisfy({ pair in
-              abs(pair.0 - pair.1) < 0.0001
-          }) else {
-        die("autoeval self-test failed: unmatched ASK did not mirror production merge semantics")
-    }
-    let collisionProfile = deterministicProfileId(2)
-    let secondCollisionProfile = deterministicProfileId(3)
-    let falseMerges = falseMergeIndicators([
-        collisionProfile: [
-            "A": Set(["meeting-1", "meeting-1", "meeting-2"]),
-            "B": ["meeting-2", "meeting-3"],
-            "C": ["meeting-3", "meeting-4"],
-            "D": ["meeting-1", "meeting-4"],
-        ],
-        secondCollisionProfile: [
-            "E": ["meeting-5"],
-            "F": ["meeting-6"],
-        ],
-    ])
-    var bucketMetrics = MutableMetrics()
-    bucketMetrics.withinMeetingFalseMergeIndicators = falseMerges.withinMeeting
-    bucketMetrics.crossMeetingFalseMergeIndicators = falseMerges.crossMeeting
-    let bucketSnapshot = bucketMetrics.snapshot()
-    guard falseMerges.withinMeeting == 4,
-          falseMerges.crossMeeting == 3,
-          bucketSnapshot.falseMergeIndicators == 7,
-          bucketSnapshot.falseMergeIndicators
-            == bucketSnapshot.withinMeetingFalseMergeIndicators
-                + bucketSnapshot.crossMeetingFalseMergeIndicators else {
-        die("autoeval self-test failed: false-merge buckets overlap or omit a truth pair")
-    }
-    print("AUTOEVAL_SELF_TEST PASS")
-}
-
-private func blendAlpha(
+func blendAlpha(
     similarity: Double,
     secondBest: Double,
     config: AutoResearchConfig
@@ -1092,7 +726,7 @@ private func blendAlpha(
     return 0
 }
 
-private func blend(_ existing: [Float], _ incoming: [Float], alpha: Float) -> [Float] {
+func blend(_ existing: [Float], _ incoming: [Float], alpha: Float) -> [Float] {
     guard existing.count == incoming.count else { return existing }
     let bounded = max(0, min(1, alpha))
     return SpeakerVectorMath.l2Normalize(zip(existing, incoming).map { old, new in
@@ -1100,7 +734,7 @@ private func blend(_ existing: [Float], _ incoming: [Float], alpha: Float) -> [F
     })
 }
 
-private func updateExemplars(
+func updateExemplars(
     state: inout SimulatedProfile,
     incoming: [Float],
     config: AutoResearchConfig
@@ -1155,14 +789,14 @@ private func mostRedundantExemplar(
 
 // MARK: - Stable identity split
 
-private func corpusFamily(_ corpus: String) -> String {
+func corpusFamily(_ corpus: String) -> String {
     if corpus.hasPrefix("ami_") { return "ami" }
     if corpus.hasPrefix("voxceleb_") { return "voxceleb" }
     if corpus.hasPrefix("voxconverse_") { return "voxconverse" }
     return corpus.split(separator: "_").first.map(String.init) ?? corpus
 }
 
-private func identitySplit(family: String, truth: String) -> ResearchSplit {
+func identitySplit(family: String, truth: String) -> ResearchSplit {
     let bucket = stableFNV1a("\(family)|\(truth)") % 10
     switch bucket {
     case 0...5: return .train
@@ -1180,7 +814,7 @@ private func stableFNV1a(_ value: String) -> UInt64 {
     return hash
 }
 
-private func deterministicProfileId(_ value: UInt64) -> UUID {
+func deterministicProfileId(_ value: UInt64) -> UUID {
     let suffix = String(format: "%012llx", value)
     return UUID(uuidString: "00000000-0000-0000-0000-\(suffix)")!
 }
